Fixed conflicts between Bootstrap and our stylesheet.
The Bootstrap reference was introduced to format the feedback messages
directed to the user's activities.
In order to solve the styling confict, the code related to the messages
formatting was located in the Bootstrap file and appended to
`config/static/styles/style.css`, thus eliminating the need for the
Bootstrap reference in the gallery index.