Photos are now associated to an user.
The Photo model has been modified to include the `user` attribute, which
is related to the users database.

The relationship was set to allow a photo to have a 'null' user, so
photos already in the database remain available if the original user has
been exlcuded from the website.

Finally, the Django Admin settings have been updated to allow filtering
photos by user, in addition to category.