Refactored HTML templates.
Created a base template for the URLs of the `users` app in order to
eliminate duplicated code. It now share the _header and _menu partials
with the `gallery` app.

All pages have the same menu now, which finally have the links to the
login and signup pages.

WARNING, this has created a bug when exhibiting messages on the index
page, which should be fixed in the next step of the project.