The apps now share the same `base.html` template.
The project has been refactored so the templates shared between
the `gallery` and `users` apps now reside in `templates/shared`. At the
moment, this means the `base.html` template.

Furthermore, the CSS style file has been updated so the use of Bootstrap
could be removed from the login and signup pages.