Implemented image editing functionality.
The image page now has an "Editar" button that allows the user to edit
the image shown.

In this context, "editing" means that the image information will be
retrieved from the database and the user will have the possibility to
modify any part of the form, including the replacement of the image
itself.

The button for deleting the image has also been added, but the
funcionality has not been implemented yet.