Modified README.md to give more accurate info.
The README.md now clarifies that the database is not part of the project
and only the images files are stored on Amazon S3.

The setup description has been slightly modified.

The next step is to add a description of how to use the application once
it's up and running.